Moved IsText from utils to text parser #1958
