+++ plaso/parsers/text_parser.py
@@ -15,7 +15,6 @@ import pyparsing
 from dfvfs.helpers import text_file
 
 from plaso.lib import errors
-from plaso.lib import utils
 from plaso.parsers import interface
 from plaso.parsers import logger
 
@@ -250,6 +249,65 @@ class PyparsingSingleLineTextParser(interface.FileObjectParser):
     # a structural fix.
     self._line_structures = self.LINE_STRUCTURES
 
+  def _IsText(self, bytes_in, encoding=None):
+    """Examine the bytes in and determine if they are indicative of text.
+  
+    Parsers need quick and at least semi reliable method of discovering whether
+    or not a particular byte stream is text or resembles text or not. This can
+    be used in text parsers to determine if a file is a text file or not for
+    instance.
+  
+    The method assumes the byte sequence is either ASCII, UTF-8, UTF-16 or method
+    supplied character encoding. Otherwise it will make the assumption the byte
+    sequence is not text, but a byte sequence.
+  
+    Args:
+      bytes_in (bytes|str): byte stream to examine.
+      encoding (Optional[str]): encoding to test, if not defined ASCII and UTF-8
+          are tried.
+  
+    Returns:
+      bool: True if the bytes stream contains text.
+    """
+    # TODO: Improve speed and accuracy of this method.
+    # Start with the assumption we are dealing with text.
+    is_text = True
+  
+    if isinstance(bytes_in, py2to3.UNICODE_TYPE):
+      return is_text
+  
+    # Check if this is ASCII text string.
+    for value in bytes_in:
+      if py2to3.PY_2:
+        value = ord(value)
+      if not 31 < value < 128:
+        is_text = False
+        break
+  
+    # We have an ASCII string.
+    if is_text:
+      return is_text
+  
+    # Check if this is UTF-8
+    try:
+      bytes_in.decode('utf-8')
+      return True
+  
+    except UnicodeDecodeError:
+      pass
+  
+    if encoding:
+      try:
+        bytes_in.decode(encoding)
+        return True
+  
+      except LookupError:
+        logger.error('Unsupported encoding: {0:s}'.format(encoding))
+      except UnicodeDecodeError:
+        pass
+  
+    return False
+
   def _ReadLine(self, text_file_object, max_len=0, depth=0):
     """Reads a line from a text file.
 
@@ -316,7 +374,7 @@ class PyparsingSingleLineTextParser(interface.FileObjectParser):
           '{2:s}]').format(
               self.MAX_LINE_LENGTH, repr(line[-10:]), self.NAME))
 
-    if not utils.IsText(line):
+    if not self._IsText(line):
       raise errors.UnableToParseFile('Not a text file, unable to proceed.')
 
     if not self.VerifyStructure(parser_mediator, line):
@@ -552,7 +610,7 @@ class PyparsingMultiLineTextParser(PyparsingSingleLineTextParser):
       raise errors.UnableToParseFile(
           'Not a text file, with error: {0!s}'.format(exception))
 
-    if not utils.IsText(self._text_reader.lines):
+    if not self._IsText(self._text_reader.lines):
       raise errors.UnableToParseFile('Not a text file, unable to proceed.')
 
     if not self.VerifyStructure(parser_mediator, self._text_reader.lines):
